Fretmarkers refactored
